Moved admin commands to server
