+++ src/FlawBOT/Modules/Server/ServerModule.cs
@@ -2,7 +2,9 @@
 using DSharpPlus.CommandsNext;
 using DSharpPlus.CommandsNext.Attributes;
 using DSharpPlus.Entities;
+using DSharpPlus.SlashCommands;
 using FlawBOT.Common;
+using FlawBOT.Properties;
 using FlawBOT.Services;
 using System;
 using System.Globalization;
@@ -94,6 +96,29 @@ namespace FlawBOT.Modules.Server
 
         #endregion COMMAND_INVITE
 
+        #region COMMAND_LEAVE
+
+        [Command("leave")]
+        [Description("Make FlawBOT leave the server.")]
+        [RequireUserPermissions(Permissions.Administrator)]
+        public async Task LeaveServer(CommandContext ctx)
+        {
+            var settings = Program.Settings;
+            await ctx.RespondAsync($"Are you sure you want {settings.Name} to leave the server?").ConfigureAwait(false);
+            var message = await ctx.RespondAsync(Resources.INFO_RESPOND).ConfigureAwait(false);
+            var interactivity = await BotServices.GetUserInteractivity(ctx, "yes", 10).ConfigureAwait(false);
+            if (interactivity.Result is null)
+            {
+                await message.ModifyAsync($"~~{message.Content}~~ {Resources.INFO_REQ_TIMEOUT}").ConfigureAwait(false);
+                return;
+            }
+
+            await BotServices.SendResponseAsync(ctx, $"Thank you for using {settings.Name}").ConfigureAwait(false);
+            await ctx.Guild.LeaveAsync().ConfigureAwait(false);
+        }
+
+        #endregion COMMAND_LEAVE
+
         #region COMMAND_PRUNE
 
         [Command("prune")]
@@ -150,6 +175,46 @@ namespace FlawBOT.Modules.Server
         }
 
         #endregion COMMAND_RENAME
+        
+        #region COMMAND_REPORT
+
+        [Hidden]
+        [Command("report")]
+        [Aliases("issue")]
+        [Description("Report a problem with FlawBOT to the developer.")]
+        public async Task ReportIssue(CommandContext ctx, [Description("Detailed description of the issue.")][RemainingText] string report)
+        {
+            if (string.IsNullOrWhiteSpace(report) || report.Length < 50)
+            {
+                await ctx.RespondAsync(Resources.ERR_REPORT_LENGTH).ConfigureAwait(false);
+                return;
+            }
+
+            await ctx.RespondAsync(Resources.INFO_REPORT_SENDER).ConfigureAwait(false);
+            var message = await ctx.RespondAsync(Resources.INFO_RESPOND).ConfigureAwait(false);
+            var interactivity = await BotServices.GetUserInteractivity(ctx, "yes", 10).ConfigureAwait(false);
+            if (interactivity.Result is null)
+            {
+                await message.ModifyAsync($"~~{message.Content}~~ {Resources.INFO_REQ_TIMEOUT}").ConfigureAwait(false);
+            }
+            else
+            {
+                var settings = Program.Settings;
+                var output = new DiscordEmbedBuilder()
+                    .WithAuthor(ctx.Guild.Owner.Username + "#" + ctx.Guild.Owner.Discriminator, iconUrl: ctx.User.AvatarUrl ?? ctx.User.DefaultAvatarUrl)
+                    .AddField("Issue", report)
+                    .AddField("Sent By", ctx.User.Username + "#" + ctx.User.Discriminator)
+                    .AddField("Server", ctx.Guild.Name + $" (ID: {ctx.Guild.Id})")
+                    .AddField("Owner", ctx.Guild.Owner.Username + "#" + ctx.Guild.Owner.Discriminator)
+                    .AddField("Confirm", $"[Click here to add this issue to GitHub]({settings.GitHubLink}/issues/new)")
+                    .WithColor(settings.DefaultColor);
+                var dm = await ctx.Member.CreateDmChannelAsync().ConfigureAwait(false);
+                await dm.SendMessageAsync(output.Build()).ConfigureAwait(false);
+                await ctx.RespondAsync("Thank You! Your report has been submitted.").ConfigureAwait(false);
+            }
+        }
+        
+        #endregion COMMAND_REPORT
 
         #region COMMAND_WARN
